adjust default job_applicants_threshold: max_applicants
LinkedIn appears to count up to 280 applicants and then stops counting (this is based on observed prints to console, where the bot repeatedly printed "280 applicants" then skipped applying to each job).

Setting the limit to 281 will ignore number of applicants by default.  It will apply to all jobs regardless of the number of previous applicants.

If a user wants to set a lower max number of applicants, they can change this on their own during their personal setup.